STEP 4. Folder drag drop.
