--- altitude/static/js/folders/event_handlers.js
@@ -1,116 +1,0 @@
-import { Const } from "../constants.js"
-import { Folder } from "../models/folder.js"
-import {
-    showErrorSnackBar,
-    showSuccessSnackBar,
-    showWarningSnackBar,
-} from "../common/snackbar.js"
-
-/**
- * folder nav warning visibility is not backend-driven, so we need to initialize the tree when it is loaded in the DOM
- */
-htmx.find("#explorer").addEventListener("htmx:load", (evt) => {
-    if (evt.target.id === "folderNavWarning") {
-        Alpine.initTree(evt.target)
-    }
-})
-
-/**
- * A folder is dragon dropped in UI BUT not yet removed on the server-side
- */
-document.body.addEventListener(Const.events.folderMoved, (event) => {
-    const movedFolderId = event.detail["movedFolderId"]
-    const movedFolder = new Folder(movedFolderId)
-    const newParentId = event.detail["newParentId"]
-    const newParent = new Folder(newParentId)
-    const oldParent = movedFolder.parent()
-
-    const msg = `Folder ${movedFolder.name()} moved from ${oldParent.name()} to ${newParent.name()}`
-    console.debug(msg)
-
-    // The server does check this, but since we remove the moved folder from the DOM, we must short-circuit
-    if (newParentId === movedFolderId) {
-        return
-    }
-
-    function handler(response) {
-        const status = response["htmx-internal-data"].xhr.status
-
-        if (status === 200) {
-            // Regardless of the target folder state, we should de-clutter the visual state of the moved folder
-            movedFolder.closeContextMenu()
-            movedFolder.clearChildren()
-
-            newParent.incrementNumOfChildren()
-            oldParent.decrementNumOfChildren()
-
-            const message = `Folder ${movedFolder.name()} moved into "${newParent.name()}"`
-            showSuccessSnackBar(message)
-
-            if (newParent.isExpanded()) {
-                // target folder is expanded, so append the moved folder
-                newParent.addChild(movedFolder)
-                movedFolder.collapse()
-            } else {
-                console.debug(
-                    "Target folder not expanded - removing the folder being moved.",
-                )
-                movedFolder.remove()
-            }
-
-            newParent.updateVisualState()
-            oldParent.updateVisualState()
-        } else if (status === 409) {
-            const message = response["htmx-internal-data"].xhr.responseText
-            showWarningSnackBar(message)
-        } else {
-            showErrorSnackBar(
-                'Error moving folder "' +
-                    movedFolder.name() +
-                    '". Status: ' +
-                    status,
-            )
-        }
-    }
-
-    console.log(event.detail)
-    htmx.ajax("put", `/htmx/folder/r/${window.ctx.getRepoId()}/move?movedFolderId=${event.detail["movedFolderId"]}&newParentId=${event.detail["newParentId"]}`, {
-        swap: "none",
-        handler: handler,
-    })
-})
-
-/**
- * A folder is added on the server
- */
-document.body.addEventListener(Const.events.folderAdded, (event) => {
-    const parentFolder = new Folder(event.detail["parentId"])
-    console.debug(
-        "Folder added event received for parent folder " + parentFolder.name(),
-    )
-
-    // Update the parent folder's number of children as we are not refreshing
-    // the tree HTML, to avoid aggressive UI changes that might confuse the user
-    parentFolder.incrementNumOfChildren()
-    parentFolder.updateVisualState()
-})
-
-/**
- * A folder is deleted on the server
- */
-document.body.addEventListener(Const.events.folderDeleted, (event) => {
-    const folder = new Folder(event.detail["id"])
-    console.debug("Folder deleted event received for folder " + folder.name())
-
-    // Update the parent folder's number of children as we are not refreshing
-    // the tree HTML, avoid aggressive UI changes that might confuse the user
-    const parent = folder.parent()
-    console.debug("Parent folder is " + parent.name())
-    parent.decrementNumOfChildren()
-    parent.updateVisualState()
-
-    showSuccessSnackBar('Folder "' + folder.name() + '" deleted')
-
-    // Without major commotion, pluck the folder from the DOM
-    folder.remove()
-})
